Rename NodePresenter to ConditionalResolver
